+++ index.html
@@ -233,7 +233,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
   <script src="js/functions.js"></script>
 </body>
 </html>
