+++ js/data.js
@@ -1,0 +1,41 @@
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Белочка1999',
+  'Елисей',
+  'Мария Викторовна',
+  'Анонимный аноним',
+  'Света',
+  'Алина',
+  'Володя'
+];
+const DESCRIPTIONS = [
+  'Фоточка, а на ней красоточка',
+  'Это я на даче, сейчас дома уже',
+  'Подумать дважды? Прости, я не думаю даже единожды',
+  'Ну и кто кому теперь не пара, Володенька?'
+];
+
+const PHOTO_COUNT = 25;
+
+const Like = {
+  MIN: 15,
+  MAX: 200
+};
+
+const MINAVATARINT = 1;
+const MAXAVATARINT = 6;
+const MINID = 1;
+const MAXCOMMENTID = 1000000000000;
+const MINCOMMENTVALUE = 1;
+const MAXCOMMENTVALUE = 10;
+
+export {MESSAGES, NAMES, DESCRIPTIONS, PHOTO_COUNT, Like, MINAVATARINT, MAXAVATARINT, MINID,
+  MAXCOMMENTID, MINCOMMENTVALUE, MAXCOMMENTVALUE };
